+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/RobotCore.java
@@ -4,8 +4,6 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
-import org.firstinspires.ftc.teamcode.robot.drive.ThreeTrackingWheelLocalizer;
 import org.firstinspires.ftc.teamcode.robot.subsystems.Subsystem;
 
 import java.util.Arrays;
