+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/AutomaticFeedforwardTuner.java
@@ -10,7 +10,6 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.RobotLog;
 
 import org.firstinspires.ftc.robotcore.internal.system.Misc;
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
 import org.firstinspires.ftc.teamcode.util.roadrunner.util.LoggingUtil;
 import org.firstinspires.ftc.teamcode.util.roadrunner.util.RegressionUtil;
