+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/FollowerPIDTuner.java
@@ -5,7 +5,6 @@ import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
 import org.firstinspires.ftc.teamcode.util.roadrunner.trajectorysequence.TrajectorySequence;
 
 /*
