+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/ManualFeedforwardTuner.java
@@ -13,8 +13,6 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.RobotLog;
 
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
-
 import java.util.Objects;
 
 import static org.firstinspires.ftc.teamcode.robot.drive.DriveConstants.MAX_ACCEL;
