+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/MaxAngularVeloTuner.java
@@ -9,8 +9,6 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
-
 import java.util.Objects;
 
 /**
