+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/MaxVelocityTuner.java
@@ -11,7 +11,6 @@ import com.qualcomm.robotcore.hardware.VoltageSensor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.teamcode.robot.drive.DriveConstants;
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
 
 import java.util.Objects;
