+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/MotorDirectionDebugger.java
@@ -7,7 +7,6 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
 
 /**
  * This is a simple teleop routine for debugging your motor configuration.
