+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/SplineTest.java
@@ -6,8 +6,6 @@ import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
-
 /*
  * This is an example of a more complex path to really test the tuning.
  */
