+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/StrafeTest.java
@@ -8,8 +8,6 @@ import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
-
 /*
  * This is a simple routine to test translational drive capabilities.
  */
