+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/TrackWidthTuner.java
@@ -11,7 +11,6 @@ import com.qualcomm.robotcore.util.MovingStatistics;
 
 import org.firstinspires.ftc.robotcore.internal.system.Misc;
 import org.firstinspires.ftc.teamcode.robot.drive.DriveConstants;
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
 
 /*
  * This routine determines the effective track width. The procedure works by executing a point turn
