+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/TrackingWheelLateralDistanceTuner.java
@@ -7,9 +7,6 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.RobotLog;
 
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
-import org.firstinspires.ftc.teamcode.robot.drive.ThreeTrackingWheelLocalizer;
-
 /**
  * Opmode designed to assist the user in tuning the `StandardTrackingWheelLocalizer`'s
  * LATERAL_DISTANCE value. The LATERAL_DISTANCE is the center-to-center distance of the parallel
