+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/drive/opmode/TurnTest.java
@@ -4,8 +4,6 @@ import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
-
 /*
  * This is a simple routine to test turning capabilities.
  */
