+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test/UltrasonicSensorTest.java
@@ -4,7 +4,6 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
-import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
 
 import java.util.List;
