+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/messages/MecanumCommandMessage.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.util.roadrunner.messages;
+package org.firstinspires.ftc.teamcode.util.messages;
 
 public final class MecanumCommandMessage {
     public long timestamp;
