--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/roadrunner/trajectorysequence/EmptySequenceException.java
@@ -1,4 +1,0 @@
-package org.firstinspires.ftc.teamcode.util.roadrunner.trajectorysequence;
-
-
-public class EmptySequenceException extends RuntimeException { }
