--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/roadrunner/trajectorysequence/sequencesegment/WaitSegment.java
@@ -1,12 +1,0 @@
-package org.firstinspires.ftc.teamcode.util.roadrunner.trajectorysequence.sequencesegment;
-
-import com.acmerobotics.roadrunner.geometry.Pose2d;
-import com.acmerobotics.roadrunner.trajectory.TrajectoryMarker;
-
-import java.util.List;
-
-public final class WaitSegment extends SequenceSegment {
-    public WaitSegment(Pose2d pose, double seconds, List<TrajectoryMarker> markers) {
-        super(seconds, pose, pose, markers);
-    }
-}
