+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/tuning/ManualFeedbackTuner.java
@@ -1,13 +1,11 @@
-package org.firstinspires.ftc.teamcode.util.roadrunner.tuning;
+package org.firstinspires.ftc.teamcode.util.tuning;
 
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
-import org.firstinspires.ftc.teamcode.ThreeDeadWheelLocalizer;
-import org.firstinspires.ftc.teamcode.TwoDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
+import org.firstinspires.ftc.teamcode.robot.drive.ThreeDeadWheelLocalizer;
 
 public final class ManualFeedbackTuner extends LinearOpMode {
     public static double DISTANCE = 64;
