+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/tuning/SplineTest.java
@@ -1,12 +1,11 @@
-package org.firstinspires.ftc.teamcode.util.roadrunner.tuning;
+package org.firstinspires.ftc.teamcode.util.tuning;
 
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
+import org.firstinspires.ftc.teamcode.robot.drive.MecanumDrive;
 
 public final class SplineTest extends LinearOpMode {
     @Override
